group user

+++ disjob-core/src/main/java/cn/ponfee/disjob/core/base/WorkerMetrics.java
@@ -54,7 +54,10 @@ public class WorkerMetrics extends ToJsonString implements Serializable {
 
         // ------------------------------------thread
 
-        private int keepAliveTime;
+        /**
+         * Worker thread keep alive time seconds
+         */
+        private long keepAliveTime;
         private int maximumPoolSize;
         private int currentPoolSize;
         private int activePoolSize;
